Update README.md
This is my first low poluy forest in blender.
If you want to download it make sure you have blender installed, or you could just look at the rendered picture.
this took me sometime to make. This was myfirst time building something in blender by myself.
I had a blast doing it. Well when I showed my friends they did complain that it looked as if it was night because of the background.
I told them " I dont know how to implement backgrounds yet".
So I hop you have fun with my project it is licencse free you can edit it as much as you want.